Fixes #13393: Rename changes-validation to change-validation
