+++ change-validation/LICENSE_EXCEPTION
