+++ change-validation/src/main/scala-templates/default/com/normation/plugins/changevalidation/EnablePluginImpl.scala
@@ -35,37 +35,13 @@
 *************************************************************************************
 */
 
-package com.normation.plugins.branding
+package com.normation.plugins.changevalidation
 
-import bootstrap.liftweb.Boot
-import com.normation.rudder.AuthorizationType.Administration
-import com.normation.rudder.domain.logger.PluginLogger
-import net.liftweb.http.ClasspathTemplates
-import net.liftweb.sitemap.Loc.LocGroup
-import net.liftweb.sitemap.Loc.Template
-import net.liftweb.sitemap.Loc.TestAccess
-import net.liftweb.sitemap.LocPath.stringToLocPath
-import net.liftweb.sitemap.Menu
-import com.normation.plugins.PluginStatus
-import com.normation.plugins.DefaultPluginDef
+import com.normation.plugins.PluginEnableImpl
 
-class BrandingPluginDef(override val status: PluginStatus) extends DefaultPluginDef {
+/*
+ * The class will be loaded by ServiceLoader, it needs an empty constructor.
+ */
 
-  override val basePackage = "com.normation.plugins.branding"
+final class CheckRudderPluginEnableImpl() extends PluginEnableImpl
 
-  def init = {}
-
-  def oneTimeInit : Unit = {}
-
-  val configFiles = Seq()
-
-  override def pluginMenuEntry: Option[Menu] = {
-    Some(Menu("brandingManagement", <span>Branding</span>) /
-        "secure" / "administration" / "brandingManagement"
-        >> LocGroup("administrationGroup")
-        >> TestAccess ( () => Boot.userIsAllowed("/secure/administration/policyServerManagement", Administration.Read) )
-        >> Template(() => ClasspathTemplates( "template" :: "brandingManagement" :: Nil ) openOr <div>Template not found</div>)
-    )
-  }
-
-}
