+++ change-validation/src/main/scala/com/normation/plugins/changevalidation/ChangeValidationPluginDef.scala
@@ -35,23 +35,21 @@
 *************************************************************************************
 */
 
-package com.normation.plugins.branding
+package com.normation.plugins.changevalidation
 
 import bootstrap.liftweb.Boot
+import com.normation.plugins._
 import com.normation.rudder.AuthorizationType.Administration
-import com.normation.rudder.domain.logger.PluginLogger
 import net.liftweb.http.ClasspathTemplates
 import net.liftweb.sitemap.Loc.LocGroup
 import net.liftweb.sitemap.Loc.Template
 import net.liftweb.sitemap.Loc.TestAccess
 import net.liftweb.sitemap.LocPath.stringToLocPath
 import net.liftweb.sitemap.Menu
-import com.normation.plugins.PluginStatus
-import com.normation.plugins.DefaultPluginDef
 
-class BrandingPluginDef(override val status: PluginStatus) extends DefaultPluginDef {
+class ChangeValidationPluginDef(override val status: PluginStatus) extends DefaultPluginDef {
 
-  override val basePackage = "com.normation.plugins.branding"
+  override val basePackage = "com.normation.plugins.changevalidation"
 
   def init = {}
 
@@ -60,11 +58,11 @@ class BrandingPluginDef(override val status: PluginStatus) extends DefaultPlugin
   val configFiles = Seq()
 
   override def pluginMenuEntry: Option[Menu] = {
-    Some(Menu("brandingManagement", <span>Branding</span>) /
-        "secure" / "administration" / "brandingManagement"
-        >> LocGroup("administrationGroup")
-        >> TestAccess ( () => Boot.userIsAllowed("/secure/administration/policyServerManagement", Administration.Read) )
-        >> Template(() => ClasspathTemplates( "template" :: "brandingManagement" :: Nil ) openOr <div>Template not found</div>)
+    Some(Menu("changeValidationManagement", <span>Change Validation</span>) /
+      "secure" / "plugins" / "changeValidationManagement"
+      >> LocGroup("pluginsGroup")
+      >> TestAccess ( () => Boot.userIsAllowed("/secure/index", Administration.Read))
+      >> Template(() => ClasspathTemplates("template" :: "ChangeValidationManagement" :: Nil ) openOr <div>Template not found</div>)
     )
   }
