--- changes-validation/src/main/resources/template/ChangesValidationManagement.html
@@ -1,94 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf8" />
-  <!-- everything here will be ignored and replaced by common-layout -->
-</head>
-<body data-lift-content-id="plugin-main">
-  <div id="plugin-main" data-lift="surround?with=common-layout;at=content">
-    <head_merge>
-      <title>Rudder :: Changes Validation</title>
-      <link rel="stylesheet" type="text/css" href="/toserve/changesvalidation/changes-validation.css" media="screen" data-lift="with-cached-resource">
-      <script src="/toserve/changesvalidation/changes-validation.js" data-lift="with-cached-resource"></script>
-    </head_merge>
-
-    <div class="lift:ChangesValidationSettings.activation tw-bs" id="workflowForm">
-      <div class="inner-portlet">
-        <div class="page-title">Change Requests (validation workflow)</div>
-        <div class="portlet-content">
-          <div class="col-xs-12 callout-fade callout-warning">
-            <div class="marker">
-              <span class="glyphicon glyphicon-info-sign"></span>
-            </div>
-            <p>
-              If enabled, all changes to configuration (Directives, Rules, Groups and Parameters) will be submitted for validation via a Change Request. <br/>
-              A new Change Request will enter the <b>Pending validation</b> status, then can be moved to <b>Pending deployment</b> (approved but not yet deployed) or <b>Deployed</b> (approved and deployed) statuses.
-            </p>
-            <p>
-              Only users with the <b>validator</b> or <b>deployer</b> roles are authorized to perform these steps (see <i><strong>/opt/rudder/etc/rudder-users.xml</strong></i>).
-            </p>
-            <p>
-              If disabled, all changes to configuration will be immediately deployed.
-            </p>
-          </div>
-          <form class="lift:form.ajax">
-            <ul>
-              <li class="rudder-form">
-                <div class="input-group">
-                  <label class="input-group-addon" for="workflowEnabled">
-                    <input id="workflowEnabled" type="checkbox">
-                    <label for="workflowEnabled" class="label-radio">
-                      <span class="ion ion-checkmark-round"></span>
-                    </label>
-                    <span class="ion ion-checkmark-round check-icon"></span>
-                  </label>
-                  <label class="form-control" for="workflowEnabled">
-                    Enable Change Requests
-                  </label>
-                </div>
-              </li>
-              <li class="rudder-form">
-                <div class="input-group">
-                  <label class="input-group-addon" for="selfVal">
-                    <input id="selfVal" type="checkbox">
-                    <label for="selfVal" class="label-radio">
-                      <span class="ion ion-checkmark-round"></span>
-                    </label>
-                    <span class="ion ion-checkmark-round check-icon"></span>
-                  </label>
-                  <label class="form-control" for="selfVal">
-                    Allow self validation
-                    <span id="selfValTooltip"></span>
-                  </label>
-                </div>
-              </li>
-              <li class="rudder-form">
-                <div class="input-group">
-                  <label class="input-group-addon" for="selfDep">
-                    <input id="selfDep" type="checkbox">
-                    <label for="selfDep" class="label-radio">
-                      <span class="ion ion-checkmark-round"></span>
-                    </label>
-                    <span class="ion ion-checkmark-round check-icon"></span>
-                  </label>
-                  <label class="form-control" for="selfDep">
-                    Allow self deployment
-                    <span id="selfDepTooltip"></span>
-                  </label>
-                </div>
-              </li>
-            </ul>
-            <lift:authz role="administration_write">
-              <input type="submit" value="Reload" id="workflowSubmit"/>
-              <span class="lift:Msg?id=updateWorkflow">[messages]</span>
-            </lift:authz>
-          </form>
-        </div>
-      </div>
-    </div>
-
-
-  </div>
-</body>
-</html>
-
